form now has functioning sliders for engineers and salary on front end for project submit
