+++ src/components/MUIReactHookForm.jsx
@@ -1,70 +1,56 @@
 /* eslint-disable max-len */
 /* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
-import { useForm, Controller, useFieldArray } from 'react-hook-form';
+import { useForm, Controller } from 'react-hook-form';
 import {
   Autocomplete,
   Box,
   Button,
-  Checkbox,
-  FormControlLabel,
   Grid,
-  Radio,
-  RadioGroup,
   TextField,
   Slider,
+  Typography,
   Select,
   MenuItem,
 } from '@mui/material';
+import MuiInput from '@mui/material/Input';
+import AttachMoneyIcon from '@mui/icons-material/AttachMoney';
+import { styled } from '@mui/material/styles';
 
-const options = ['A', 'B', 'C', 'D'];
-const objOptions = [
-  { value: 65, label: 'A' },
-  { value: 66, label: 'B' },
-  { value: 67, label: 'C' },
+const Input = styled(MuiInput)`
+  width: 42px;
+`;
+
+// const options = ['Finance', 'Human Resources', 'Market Research', 'Industry'];
+const skillOptions = [
+  { value: 1, label: 'JavaScript' },
+  { value: 2, label: 'React.js' },
+  { value: 3, label: 'Python' },
 ];
-const myHelper = {
-  email: {
-    required: 'Email is Required',
-    pattern: 'Invalid Email Address',
-  },
-};
 
 export default function MUIReactHookForm() {
   const { control, handleSubmit } = useForm({
     reValidateMode: 'onBlur',
   });
-  const {
-    fields: members,
-    append: appendMemberRow,
-    remove: removeMemberRow,
-  } = useFieldArray({
-    control,
-    name: 'members',
-  });
-
-  console.count('app rerender');
 
   const handleOnSubmit = (evt) => {
     console.log(evt);
   };
 
-  const addNewMember = () => appendMemberRow({ email: '', role: 'user' });
-
   return (
     <Box component="form" onSubmit={handleSubmit(handleOnSubmit)}>
       <Grid container spacing={3}>
-        <Grid item xs={6}>
+        <Grid item xs={12}>
           <Controller
             control={control}
-            name="username"
+            name="name"
             defaultValue=""
             render={({ field }) => (
               <TextField
                 {...field}
+                variant="standard"
                 fullWidth
-                variant="filled"
-                label="username"
+                label="What is the project's name?"
               />
             )}
           />
@@ -72,52 +58,127 @@ export default function MUIReactHookForm() {
         <Grid item xs={6}>
           <Controller
             control={control}
-            name="selection"
+            name="industryId"
             defaultValue=""
             render={({ field }) => (
-              <Select {...field} fullWidth>
-                <MenuItem value="A">Select 1</MenuItem>
-                <MenuItem value="B">Select 2</MenuItem>
-                <MenuItem value="C">Select 3</MenuItem>
-                <MenuItem value="D">Select 4</MenuItem>
-              </Select>
+              <>
+                <Typography>
+                  Client&apos;s Industry:
+                </Typography>
+                <Select
+                  {...field}
+                  fullWidth
+                >
+                  <MenuItem value="1">Airline</MenuItem>
+                  <MenuItem value="2">Finance</MenuItem>
+                  <MenuItem value="3">Market Research</MenuItem>
+                  <MenuItem value="4">Human Resources</MenuItem>
+                  <MenuItem value="5">Technology</MenuItem>
+                </Select>
+              </>
+
             )}
           />
         </Grid>
-        <Grid item xs={6}>
+        <Grid item xs={12}>
           <Controller
             control={control}
-            name="auto-complete"
-            defaultValue={options[0]}
-            render={({ field: { ref, onChange, ...field } }) => (
-              <Autocomplete
-                options={options}
-                onChange={(_, data) => onChange(data.value)}
-                defaultValue={options[0]}
-                renderInput={(params) => (
-                  <TextField
-                    {...params}
-                    {...field}
-                    fullWidth
-                    inputRef={ref}
-                    variant="filled"
-                    label="Auto-Complete"
-                  />
-                )}
+            name="summary"
+            defaultValue="This client needs a ..."
+            render={({ field }) => (
+              <TextField
+                {...field}
+                id="outlined-multiline-static"
+                label="Summarize the project's requirements here:"
+                multiline
+                fullWidth
+                rows={5}
               />
             )}
           />
         </Grid>
+        <Grid item xs={12}>
+
+          <Controller
+            control={control}
+            name="numberEngineers"
+            defaultValue={1}
+            render={({ field: { value, ...field } }) => (
+              <>
+                <Typography>
+                  Number of Engineers required:
+                </Typography>
+                <Slider
+                  {...field}
+                  marks
+                  max={10}
+                  min={1}
+                  step={1}
+                  value={value}
+                  valueLabelDisplay="auto"
+                />
+
+              </>
+
+            )}
+          />
+        </Grid>
+        <Grid item xs={12}>
+
+          <Controller
+            control={control}
+            name="minimumSalary"
+            defaultValue={30}
+            render={({ field: { value, ...field } }) => (
+              <>
+                <Typography id="input-slider" gutterBottom>
+                  Minimum Salary Paid per Hour:
+                </Typography>
+                <Grid container spacing={2} alignItems="center">
+                  <Grid item>
+                    <AttachMoneyIcon />
+                  </Grid>
+                  <Grid item xs>
+                    <Slider
+                      {...field}
+                      value={typeof value === 'number' ? value : 0}
+                      // onChange={handleSliderChange}
+                      aria-labelledby="input-slider"
+                    />
+                  </Grid>
+                  <Grid item>
+                    <Input
+                      {...field}
+                      value={value}
+                      size="small"
+                      // onChange={handleInputChange}
+                      // onBlur={handleBlur}
+                      inputProps={{
+                        step: 10,
+                        min: 0,
+                        max: 100,
+                        type: 'number',
+                        'aria-labelledby': 'input-slider',
+                      }}
+                    />
+                  </Grid>
+                </Grid>
+
+              </>
+
+            )}
+          />
+        </Grid>
         <Grid item xs={6}>
           <Controller
             control={control}
-            name="object-component"
-            defaultValue={[objOptions[0]]}
+            name="projectSkills"
+            defaultValue={[skillOptions[0]]}
             render={({ field: { ref, onChange, ...field } }) => (
               <Autocomplete
                 multiple
-                options={objOptions}
-                defaultValue={[objOptions[0]]}
+                options={skillOptions}
+                defaultValue={[skillOptions[0]]}
                 getOptionLabel={(option) => option.label}
                 onChange={(_, data) => onChange(data)}
                 renderInput={(params) => (
@@ -126,162 +187,15 @@ export default function MUIReactHookForm() {
                     {...params}
                     fullWidth
                     inputRef={ref}
-                    variant="filled"
-                    label="object-complete"
+                    variant="standard"
+                    label="Skills Required:"
                   />
                 )}
               />
             )}
           />
         </Grid>
-        <Grid item xs={6}>
-          <Controller
-            control={control}
-            name="check-box"
-            defaultValue={false}
-            render={({ field: { value, onChange, ...field } }) => (
-              <FormControlLabel
-                control={
-                  <Checkbox onChange={onChange} checked={value} {...field} />
-                  }
-                label="checkbox"
-              />
-            )}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <Controller
-            control={control}
-            name="radio-group"
-            defaultValue=""
-            render={({ field }) => (
-              <RadioGroup {...field}>
-                <FormControlLabel
-                  value="choice-1"
-                  control={<Radio />}
-                  label="A"
-                />
-                <FormControlLabel
-                  value="choice-2"
-                  control={<Radio />}
-                  label="B"
-                />
-                <FormControlLabel
-                  value="choice-3"
-                  control={<Radio />}
-                  label="C"
-                />
-                <FormControlLabel
-                  value="choice-4"
-                  control={<Radio />}
-                  label="D"
-                />
-              </RadioGroup>
-            )}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <Controller
-            control={control}
-            name="slider"
-            defaultValue={50}
-            render={({ field: { value, ...field } }) => (
-              <Slider
-                {...field}
-                marks
-                max={100}
-                min={0}
-                step={5}
-                value={value}
-              />
-            )}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <Controller
-            control={control}
-            name="slider-range"
-            defaultValue={[0, 50]}
-            render={({ field: { value, ...field } }) => (
-              <Slider
-                {...field}
-                marks
-                max={100}
-                min={0}
-                step={5}
-                value={value}
-              />
-            )}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <Controller
-            control={control}
-            name="email"
-            defaultValue=""
-            rules={{
-              required: true,
-              pattern: /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/,
-            }}
-            render={({ field, fieldState: { error } }) => (
-              <TextField
-                {...field}
-                type="email"
-                fullWidth
-                label="Email With Validation"
-                error={error !== undefined}
-                helperText={error ? myHelper.email[error.type] : ''}
-              />
-            )}
-          />
-        </Grid>
-        <Grid item xs={12}>
-          {members.map((field, index) => (
-            <Grid container key={field.id} spacing={1} alignItems="center">
-              <Grid item xs={6}>
-                <Controller
-                  control={control}
-                    // must use . for the object key!!!
-                  name={`members.${index}.email`}
-                  defaultValue=""
-                  render={({ field }) => (
-                    <TextField {...field} type="email" fullWidth />
-                  )}
-                />
-              </Grid>
-              <Grid item xs={4}>
-                <Controller
-                  control={control}
-                    // must use . for the object key!!!
-                  name={`members.${index}.role`}
-                  defaultValue="user"
-                  render={({ field }) => (
-                    <Select {...field} fullWidth>
-                      <MenuItem value="user">Member</MenuItem>
-                      <MenuItem value="admin">Admin</MenuItem>
-                    </Select>
-                  )}
-                />
-              </Grid>
-              <Grid item xs={2}>
-                <Button
-                  color="error"
-                  variant="text"
-                  onClick={() => removeMemberRow(index)}
-                >
-                  Delete
-                </Button>
-              </Grid>
-            </Grid>
-          ))}
-          <Grid item xs={12}>
-            Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et. Cras mattis consectetur purus sit amet fermentum. Cras justo odio, dapibus ac facilisis in, egestas eget quam. Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Praesent commodo cursus magna, vel scelerisque nisl consectetur et.
 
-          </Grid>
-          <Button variant="contained" onClick={addNewMember}>
-            Add
-          </Button>
-        </Grid>
         <Grid item xs={12}>
           <Button type="submit">Submit</Button>
         </Grid>
